pyproject.toml
[project]
name = "multi-agent-research-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "dspy",
    "openai",
    "jupyter",
    "pydantic",
    "brave-search-python-client",
    "nest-asyncio>=1.6.0",
    "pytest",
    "langfuse",
    "openinference-instrumentation-dspy",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
